refactor(secure): move passwordChangedAt to BaseUser
passwordChangedAt is consumed only by SessionInvalidationListener (a bundle
class), so it belongs on the bundle's BaseUser mapped superclass rather than
the app's BSUser entity. Retype the two bundle-side instanceof checks to
BaseUser; the app controller keeps BSUser since it legitimately references its
own concrete entity. No schema change: mapped-superclass columns inline into
the same sauvabelin_netbs_users table.
